requirements.txt
# Dash and related
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0

# Data Processing
pandas==1.5.1
numpy==1.24.3
pandas_ta
matplotlib
statsmodels
arch

# APIs and Data Sources
yfinance
MetaTrader5
fredapi
requests
beautifulsoup4
wikipedia
arxiv
stackapi

# Machine Learning and AI
openai
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
longchain-experimental==0.0.58
langchainhub==0.1.15
faiss-cpu

# Environment and Utilities
python-dotenv
pydantic
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
urllib3==1.26.15
websocket==0.2.1
websocket-client==0.44.0
Werkzeug==2.3.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0

# Report Generation
reportlab

# Development
ipykernel
